Move function test to correct file
